move BaseChanceEntry up into ChancedOutput improve toString()
