+++ src/main/java/gregtech/api/recipes/chance/output/BoostableChanceOutput.java
@@ -33,6 +33,7 @@ public abstract class BoostableChanceOutput<T> extends ChancedOutput<T> implemen
         return "BoostableChanceOutput{" +
                 "ingredient=" + getIngredient() +
                 ", chance=" + getChance() +
+                ", maxChance=" + getMaxChance() +
                 ", chanceBoost=" + getChanceBoost() +
                 '}';
     }
