+++ src/main/java/gregtech/api/recipes/chance/output/ChancedOutput.java
@@ -1,6 +1,6 @@
 package gregtech.api.recipes.chance.output;
 
-import gregtech.api.recipes.chance.BaseChanceEntry;
+import gregtech.api.recipes.chance.ChanceEntry;
 
 import org.jetbrains.annotations.NotNull;
 
@@ -9,14 +9,42 @@ import org.jetbrains.annotations.NotNull;
  *
  * @param <T> the type of ingredient contained by the output
  */
-public abstract class ChancedOutput<T> extends BaseChanceEntry<T> {
+public abstract class ChancedOutput<T> implements ChanceEntry<T> {
+
+    private final T ingredient;
+    private int chance;
+    private int maxChance;
 
     public ChancedOutput(@NotNull T ingredient, int chance) {
-        super(ingredient, chance);
+        this.ingredient = ingredient;
+        this.chance = chance;
+        this.maxChance = ChancedOutputLogic.getMaxChancedValue();
     }
 
 
     public ChancedOutput(@NotNull T ingredient, int chance, int maxChance) {
-        super(ingredient, chance, maxChance);
+        this.ingredient = ingredient;
+        this.chance = chance;
+        this.maxChance = maxChance;
+    }
+
+    @Override
+    public @NotNull T getIngredient() {
+        return ingredient;
+    }
+
+    @Override
+    public int getChance() {
+        return chance;
+    }
+
+    @Override
+    public int getMaxChance() {
+        return maxChance;
+    }
+
+    public void scale(int scalar) {
+        this.chance *= scalar;
+        this.maxChance *= scalar;
     }
 }
